requirements.txt
pyodbc
python-ldap